Backend/requirements.txt
fastapi[standard]
aiohttp
python-dotenv
transformers
torch
unsloth